If we can't open httpd.conf, print a warning, but don't die. Otherwise, you can't install a module into a package root.
git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@84238 13f79535-47bb-0310-9956-ffa450edef68

+++ src/support/apxs.pl
@@ -510,15 +510,18 @@ if ($opt_i or $opt_e) {
             } else {
                  $content =~ s|^(.*\n)#?\s*$amd[^\n]*\n|$1$c$amd\n|sg;
             }
-        }
+	}
         if (@lmd or @amd) {
-            open(FP, ">$CFG_SYSCONFDIR/$CFG_TARGET.conf.new") || die;
-            print FP $content;
-            close(FP);
-            system("cp $CFG_SYSCONFDIR/$CFG_TARGET.conf $CFG_SYSCONFDIR/$CFG_TARGET.conf.bak && " .
-                   "cp $CFG_SYSCONFDIR/$CFG_TARGET.conf.new $CFG_SYSCONFDIR/$CFG_TARGET.conf && " .
-                   "rm $CFG_SYSCONFDIR/$CFG_TARGET.conf.new");
-        }
+            if (open(FP, ">$CFG_SYSCONFDIR/$CFG_TARGET.conf.new")) {
+                print FP $content;
+                close(FP);
+                system("cp $CFG_SYSCONFDIR/$CFG_TARGET.conf $CFG_SYSCONFDIR/$CFG_TARGET.conf.bak && " .
+                       "cp $CFG_SYSCONFDIR/$CFG_TARGET.conf.new $CFG_SYSCONFDIR/$CFG_TARGET.conf && " .
+                       "rm $CFG_SYSCONFDIR/$CFG_TARGET.conf.new");
+            } else {
+                print STDERR "unable to open configuration file\n";
+            }
+	}
     }
 }
